+++ Assets/Polaris - Low Poly Ecosystem/Polaris - Low Poly Terrain Engine/Runtime/Scripts/Compression/LzfLibrary/LZF.cs
@@ -100,7 +100,7 @@ namespace Lzf
                     hval = (hval << 8) | input[iidx + 2];
                     hslot = ((hval ^ (hval << 5)) >> (int)(((3 * 8 - HLOG)) - hval * 5) & (HSIZE - 1));
                     reference = HashTable[hslot];
-                    HashTable[hslot] = (long)iidx;
+                    HashTable[hslot] = iidx;
 
 
                     if ((off = iidx - reference - 1) < MAX_OFF
